+++ Asteroids/Entity.h
@@ -46,7 +46,7 @@ class Entity
 
     int      asteroid_level_;
     float    scale_factor_;
-    unsigned game_level_;
+    float    game_level_;
 
     MovementComponent*  movement_component_;
     AnimationComponent* animation_component_;
